+++ Package.swift
@@ -46,6 +46,10 @@ let package = Package(
             ],
             resources: [
                 .process("Resources")
+            ],
+            linkerSettings: [
+                .linkedFramework("CoreWLAN"),
+                .linkedFramework("CoreLocation")
             ]
         ),
         .testTarget(